inertial scrolling working! some drags are unresponsive, but I can't tell if that's just my simulator or an actual problem

+++ source/main.h
@@ -40,7 +40,7 @@ Player* opponentPlayer = NULL;
 #define MAX_TOUCHES 10
 CTouch touches[MAX_TOUCHES];
 
-int32 worldScrollSpeed = 0;
+float worldScrollSpeed = 0;
 
 float getAngleDiff(int32 x0, int32 y0, int32 x1, int32 y1);
 float getAngleDiff(CTouch* touch);
